refactor: simplify session date components and inline tab navigation
