--- components/Session/Date/Tab.vue
@@ -1,15 +1,0 @@
-<template>
-  <nav :class="$style.tab">
-    <slot />
-  </nav>
-</template>
-
-<style module>
-.tab {
-  display: flex;
-  justify-content: center;
-  gap: 0.5rem;
-  font-family: Inter, sans-serif;
-  color: #fff;
-}
-</style>
